Add files via upload
Finalised;
upload and update functionality works.
correct tables are updated.
only allowed files are uploaded. However because of how the for loop works, I couldn't figure out a way to display individual errors for files so instead the not-allowed files are simply not allowed to be uploaded and I've specified on the webpage which ext's are safe to upload.